+++ wemake_python_styleguide/constants.py
@@ -198,8 +198,8 @@ MAX_LEN_YIELD_TUPLE: Final = 5
 # Used as a default filename, when it is not passed by flake8:
 STDIN: Final = 'stdin'
 
-# Used as a special name for unused variables:
-UNUSED_VARIABLE: Final = '_'
+# Used as a special name patterns for unused variables, like _, __:
+UNUSED_VARIABLE_REGEX: Final = re.compile(r'^_+$')
 
 # Used to specify as a placeholder for `__init__`:
 INIT: Final = '__init__'